fix: delete dto package and move request, response package to api package
