+++ src/main/java/com/takemissinghome/weakperson/api/request/BenefitDataRequest.java
@@ -1,4 +1,4 @@
-package com.takemissinghome.weakperson.api.dto.request;
+package com.takemissinghome.weakperson.api.request;
 
 import com.takemissinghome.gov.response.Content;
 import lombok.Getter;
